fix remarks

+++ ErshovVladislav/Task4/Task4/src/main/java/ershov/ExamSystemController.java
@@ -32,7 +32,7 @@ public class ExamSystemController {
 	}
 
 	@RequestMapping("/count")
-	public long examSystemCount() {
+	public int examSystemCount() {
 		return examSystem.count();
 	}
